+++ windows/src/developer/kmcmpdll/kmcmpdll.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <windows.h>
+#include "compfile.h"
 
 BOOL AddCompileString(LPSTR buf);
 BOOL AddCompileMessage(DWORD msg);
@@ -12,3 +13,4 @@ extern BOOL FWarnDeprecatedCode;
 extern int currentLine;
 
 PWSTR strtowstr(PSTR in);
+PFILE_STORE FindSystemStore(PFILE_KEYBOARD fk, DWORD dwSystemID);
